requirements.txt
opencv-python~=4.8.0.74
numpy~=1.24.4
torch~=2.3.0
transformers~=4.42.3
scikit-learn~=1.3.0
torchvision~=0.18.0
natsort~=8.4.0
clip~=0.2.0
Pillow~=10.0.1
Levenshtein
fuzzywuzzy~=0.18.0
lxml~=4.9.3
scipy~=1.9.3
Shapely~=1.8.0
arabic-reshaper~=3.0.0
brnolm~=0.3.0
numba~=0.57.1
scikit-image~=0.21.0
lmdb~=1.4.1
python-dateutil~=2.8.2
matplotlib~=3.8.0